agent/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.32.0
python-dotenv>=1.0.1
lxml>=5.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
